Review commits and code style

+++ src/accelerate/test_utils/scripts/test_script.py
@@ -401,7 +401,7 @@ def training_check():
         from accelerate.utils.dataclasses import IntelPyTorchExtensionPlugin
 
         AcceleratorState._reset_state()
-        ipex_plugin = IntelPyTorchExtensionPlugin(use_ipex=True, dtype=torch.bfloat16)
+        ipex_plugin = IntelPyTorchExtensionPlugin(use_ipex=True)
         accelerator = Accelerator(mixed_precision="bf16", cpu=True, ipex_plugin=ipex_plugin)
         train_dl = DataLoader(train_set, batch_size=batch_size, shuffle=True, generator=generator)
         model = RegressionModel()
@@ -428,7 +428,7 @@ def training_check():
         from accelerate.utils.dataclasses import XPUPlugin
 
         AcceleratorState._reset_state()
-        xpu_plugin = XPUPlugin(use_xpu=True, dtype=torch.bfloat16)
+        xpu_plugin = XPUPlugin(use_xpu=True)
         accelerator = Accelerator(mixed_precision="bf16", cpu=False, xpu_plugin=xpu_plugin)
         train_dl = DataLoader(train_set, batch_size=batch_size, shuffle=True, generator=generator)
         model = RegressionModel()
